Moved RelationshipState to user
