+++ app/src/main/resources/swagger/api-docs.json
@@ -709,6 +709,7 @@
             }
           }
         },
+        "deprecated" : true,
         "security" : [ {
           "bearerAuth" : [ "global" ]
         } ]
@@ -2898,27 +2899,18 @@
         }
       },
       "ProductsResource" : {
-        "title": "ProductsResource",
-        "required": [
-          "authorized",
-          "description",
-          "id",
-          "imageUrl",
-          "logo",
-          "productOnBoardingStatus",
-          "title",
-          "urlBO"
-        ],
-        "type": "object",
-        "properties": {
-          "activatedAt": {
-            "type": "string",
-            "description": "Date the products was activated",
-            "format": "date-time"
+        "title" : "ProductsResource",
+        "required" : [ "authorized", "description", "id", "imageUrl", "logo", "productOnBoardingStatus", "title", "urlBO" ],
+        "type" : "object",
+        "properties" : {
+          "activatedAt" : {
+            "type" : "string",
+            "description" : "Date the products was activated",
+            "format" : "date-time"
           },
-          "authorized": {
-            "type": "boolean",
-            "description": "flag indicating whether the logged user has the authorization to manage the product",
+          "authorized" : {
+            "type" : "boolean",
+            "description" : "flag indicating whether the logged user has the authorization to manage the product",
             "example" : false
           },
           "backOfficeEnvironmentConfigurations" : {
@@ -2956,14 +2948,10 @@
             "type" : "string",
             "description" : "Product logo's background color"
           },
-          "productOnBoardingStatus": {
-            "type": "string",
-            "description": "Product's onBoarding status",
-            "enum": [
-              "ACTIVE",
-              "INACTIVE",
-              "PENDING"
-            ]
+          "productOnBoardingStatus" : {
+            "type" : "string",
+            "description" : "Product's onBoarding status",
+            "enum" : [ "ACTIVE", "INACTIVE", "PENDING" ]
           },
           "title" : {
             "type" : "string",
@@ -3023,26 +3011,18 @@
         }
       },
       "SubProductResource" : {
-        "title": "SubProductResource",
-        "required": [
-          "id",
-          "productOnBoardingStatus",
-          "title"
-        ],
-        "type": "object",
-        "properties": {
-          "id": {
-            "type": "string",
-            "description": "Product's unique identifier"
+        "title" : "SubProductResource",
+        "required" : [ "id", "productOnBoardingStatus", "title" ],
+        "type" : "object",
+        "properties" : {
+          "id" : {
+            "type" : "string",
+            "description" : "Product's unique identifier"
           },
-          "productOnBoardingStatus": {
-            "type": "string",
-            "description": "Product's onBoarding status",
-            "enum": [
-              "ACTIVE",
-              "INACTIVE",
-              "PENDING"
-            ]
+          "productOnBoardingStatus" : {
+            "type" : "string",
+            "description" : "Product's onBoarding status",
+            "enum" : [ "ACTIVE", "INACTIVE", "PENDING" ]
           },
           "title" : {
             "type" : "string",
