+++ connector-api/src/main/java/it/pagopa/selfcare/dashboard/connector/model/institution/InstitutionInfo.java
@@ -1,5 +1,6 @@
 package it.pagopa.selfcare.dashboard.connector.model.institution;
 
+import it.pagopa.selfcare.dashboard.connector.model.user.RelationshipState;
 import lombok.Data;
 
 import java.util.Objects;
